prob02-danny539

+++ prob02/main.cpp
@@ -3,6 +3,7 @@
 #include <iostream>
 
 int main()
+
 {
   // one gallon of paint covers 340 square feet
   int coverage = 340;
@@ -11,10 +12,11 @@ int main()
 
   // calculate how many gallons of paint it will take to cover
   // the area with 2 coats of paint
-  int area=height*width*2;
-  int gallons = area / coverage;
+  double area = height*width*2;
+  double gallons = area / coverage;
 
   std::cout << "It will take " << gallons << " gallons of paint.";
 
   return 0;
+
 }
